requirements.txt
openai
pygame
realtimestt
googlesearch-python
python-dotenv